Parse BLAST results

+++ scripts/BLAST.py
@@ -8,14 +8,12 @@ import urllib2
 import pickle
 import os.path
 
-def blast(protein_id):
-    #first try to find a cached result
-    if (os.path.exists(os.path.join('cache', protein_id + '_BLAST.pkl'))):
-        f = open(os.path.join('cache', protein_id + '_BLAST.pkl'), 'rb')
-        p = pickle.load(f)
+def query_server(protein_id):
+    if (os.path.exists(protein_id + '_BLAST.html')):
+        f = open(protein_id + '_BLAST.html')
+        result = f.read()
         f.close()
-        return p
-    
+        return result
     baseUrl = 'http://blast.ncbi.nlm.nih.gov/Blast.cgi?CMD=Put&QUERY='
     url = baseUrl + protein_id + '&DATABASE=swissprot&PROGRAM=blastp&FILTER=T&EXPECT=100&FORMAT_TYPE=Text'
     fh = urllib2.urlopen(url)
@@ -59,21 +57,50 @@ def blast(protein_id):
     o = open(output, 'w') 
     o.write(result) 
     o.close()
-    
+
+def parse_id(s):
+    return re.split("\|", s)[3]
+
+def parse_ids(s):
+    return [parse_id(id_s) for id_s in re.split(";", s)]
+
+def parse_blast_result(s):
+    lines = re.split("\n", s)
+    blast_hits = []
+    for line in lines:
+        if len(line) < 1:
+            #this line is empty
+            continue
+        if line[0] == '#':
+            #this line is a comment
+            continue
+        cols = re.split("\t", line)
+        h = {'query' : parse_id(cols[0]),
+             'subjects' : parse_ids(cols[1]),
+             'evalue' : float(cols[11]),
+             'bit-score' : float(cols[12])}
+        blast_hits.append(h)
+    return blast_hits
+        
+def blast(protein_id):
+    #first try to find a cached result
+    if (os.path.exists(os.path.join('cache', protein_id + '_BLAST.pkl'))):
+        f = open(os.path.join('cache', protein_id + '_BLAST.pkl'), 'rb')
+        p = pickle.load(f)
+        f.close()
+        return p
+    result = query_server(protein_id)
     #Find the result between the PRE tags
     m = re.search("^<PRE>(.*)^</PRE>", result, re.MULTILINE + re.DOTALL)
-    r = m.groups()[0]
-    
+    r = parse_blast_result(m.groups()[0])
     #write to file system cache
     f = open(os.path.join('cache', protein_id + '_BLAST.pkl'), 'wb')
     pickle.dump(r, f)
     f.close()
-    
     return r
 
 if __name__ == "__main__":
     f = open('proteins.txt', 'r') 
     for line in f:
         protein_id = line.strip("\n")
-        blast(protein_id) 
- 
+        print blast(protein_id)[1]
